Fix warning messages in GSM and speex because of previous update to GSM patchlevel 10 and speex 1.1.12
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@523 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjmedia/src/pjmedia-codec/speex/kiss_fft.c
@@ -87,7 +87,7 @@ static void kf_bfly4(
     if (!st->inverse) {
        int i;
        kiss_fft_cpx *x=Fout;
-       for (i=0;i<4*m;i++)
+       for (i=0;i<(int)(4*m);i++)
        {
           x[i].r = PSHR16(x[i].r,2);
           x[i].i = PSHR16(x[i].i,2);
